+++ docs/changelog.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add components init command
 - Add storage config and config parser
 - Create base environment to run commands
+- Add commands redaction 
 
 ### Changed